(refactor) : the middleware dir. - organise the file struct for the middleware dir.
