+++ src/routes/users.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import { getUsers } from '@Controllers/usersController';
-import verifyUserToken from '@Middlewares/verifyUserToken';
+import { verifyUserToken } from '@Middlewares/authMiddlewares/verifyUserToken';
 
 const router = Router();
